@@ -1,12 +1,12 @@
 {
   "name": "@lab900/ui",
-  "version": "17.0.5",
+  "version": "18.0.3",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
   "licens": "MIT",
   "peerDependencies": {
-    "@angular/common": ">=17.0.0",
-    "@angular/core": ">=17.0.0",
-    "@angular/material": ">=17.0.0",
+    "@angular/common": ">=18.1.0",
+    "@angular/core": ">=18.1.0",
+    "@angular/material": ">=18.1.0",
     "@ngx-translate/core": ">=15.0.0"
   },
   "repository": {
