@@ -1,6 +1,6 @@
 {
   "name": "angular-library-ui",
-  "version": "17.0.0",
+  "version": "18.0.0",
   "repository": "https://github.com/lab900/angular-library-ui",
   "homepage": "https://github.com/lab900/angular-library-ui",
   "author": "Lab900 <info@lab900.com> (https://lab900.com)",
@@ -17,48 +17,41 @@
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "pretty:check": "prettier -c \"{src,lib}/**/**.{ts,js}\"",
-    "pretty:fix": "prettier  --write \"{src,lib}/**/**.{ts,js}\"",
-    "pretty:html:fix": "prettier --no-config --print-width=140 --write \"{src,lib}/**/**.html\"",
+    "prettier": "prettier --write .",
     "clean": "rm -rf node_modules/ && npm install"
   },
   "dependencies": {
-    "@angular/animations": "^17.3.5",
-    "@angular/cdk": "^17.3.5",
-    "@angular/common": "^17.3.5",
-    "@angular/compiler": "^17.3.5",
-    "@angular/core": "^17.3.5",
-    "@angular/forms": "^17.3.5",
-    "@angular/material": "^17.3.5",
-    "@angular/platform-browser": "^17.3.5",
-    "@angular/platform-browser-dynamic": "^17.3.5",
-    "@angular/router": "^17.3.5",
+    "@angular/animations": "^18.1.1",
+    "@angular/cdk": "^18.1.1",
+    "@angular/common": "^18.1.1",
+    "@angular/compiler": "^18.1.1",
+    "@angular/core": "^18.1.1",
+    "@angular/forms": "^18.1.1",
+    "@angular/material": "^18.1.1",
+    "@angular/platform-browser": "^18.1.1",
+    "@angular/platform-browser-dynamic": "^18.1.1",
+    "@angular/router": "^18.1.1",
     "@ngx-translate/core": "^15.0.0",
-    "ngx-markdown": "^17.2.0",
+    "ngx-markdown": "^18.0.0",
     "prismjs": "^1.29.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.0.0",
     "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.5",
-    "@angular-eslint/builder": "17.3.0",
-    "@angular-eslint/eslint-plugin": "17.3.0",
-    "@angular-eslint/eslint-plugin-template": "17.3.0",
-    "@angular-eslint/schematics": "17.3.0",
-    "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "^17.3.5",
-    "@angular/compiler-cli": "^17.3.5",
-    "@angular/language-service": "^17.3.5",
+    "@angular-devkit/build-angular": "^18.1.1",
+    "@angular/cli": "^18.1.1",
+    "@angular/compiler-cli": "^18.1.1",
+    "@eslint/js": "^9.8.0",
+    "@types/eslint__js": "^8.42.3",
     "@types/jasmine": "~3.10.2",
     "@types/jasminewd2": "~2.0.10",
     "@types/node": "^20.11.1",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
     "angular-cli-ghpages": "^2.0.0-beta.2",
-    "eslint": "^8.57.0",
+    "angular-eslint": "18.1.0",
+    "eslint": "^9.8.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^5.2.1",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.4.1",
@@ -66,10 +59,10 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^17.3.0",
-    "prettier": "3.2.5",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "typescript": "~5.4.5"
+    "ng-packagr": "^18.1.0",
+    "prettier": "^3.3.3",
+    "prettier-eslint": "^16.3.0",
+    "typescript": "~5.4.5",
+    "typescript-eslint": "^8.0.0-alpha.58"
   }
 }
