@@ -1,9 +1,4 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  ViewEncapsulation,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, input, ViewEncapsulation } from '@angular/core';
 import { Icon } from '../../models/nav-item.model';
 import { MatIconModule } from '@angular/material/icon';
 
@@ -12,16 +7,15 @@ import { MatIconModule } from '@angular/material/icon';
   styleUrls: ['./icon.component.scss'],
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  template: ` @if (icon?.name) {
-      <mat-icon>{{ icon.name }}</mat-icon>
+  template: ` @if (icon()?.name) {
+      <mat-icon>{{ icon().name }}</mat-icon>
     }
-    @if (icon?.svgName) {
-      <mat-icon [svgIcon]="icon.svgName"></mat-icon>
+    @if (icon()?.svgName) {
+      <mat-icon [svgIcon]="icon().svgName" />
     }`,
   imports: [MatIconModule],
   standalone: true,
 })
 export class IconComponent {
-  @Input()
-  public icon: Icon;
+  public readonly icon = input.required<Icon>();
 }
