@@ -1,17 +1,8 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  OnInit,
-  ViewEncapsulation,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, effect, inject, input, model, ViewEncapsulation } from '@angular/core';
 import { IsActiveMatchOptions, NavigationEnd, Router } from '@angular/router';
 import { NavItem } from '../../models/nav-item.model';
-import { BehaviorSubject, Observable } from 'rxjs';
-import { SubscriptionBasedDirective } from '../../../common/directives/subscription-based.directive';
 import { filter } from 'rxjs/operators';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
-import { AsyncPipe } from '@angular/common';
 import { NavItemButtonComponent } from '../nav-item-button/nav-item-button.component';
 
 @Component({
@@ -21,80 +12,43 @@ import { NavItemButtonComponent } from '../nav-item-button/nav-item-button.compo
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
-  imports: [NavItemButtonComponent, AsyncPipe],
+  imports: [NavItemButtonComponent],
 })
-export class NavItemComponent
-  extends SubscriptionBasedDirective
-  implements OnInit
-{
-  @Input({ required: true })
-  public readonly item: NavItem;
+export class NavItemComponent {
+  public readonly router = inject(Router);
+  public readonly breakpointObserver = inject(BreakpointObserver);
 
-  @Input()
-  public readonly indentLevels = true;
+  public readonly item = input.required<NavItem>();
+  public readonly indentLevels = input<boolean>(true);
+  public readonly showLevelArrows = input<boolean>(false);
+  public readonly depth = input<number>(0);
+  public readonly allowOverlayMenuUntil = input<string | string[]>(Breakpoints.XSmall);
+  public readonly routeMatchOptions = input<IsActiveMatchOptions | { exact: boolean } | undefined>(undefined);
 
-  @Input()
-  public readonly showLevelArrows = false;
+  public readonly disabled = model<boolean>(false);
+  public readonly expanded = model<boolean>(false);
 
-  @Input()
-  public readonly depth = 0;
-
-  @Input()
-  public readonly allowOverlayMenuUntil: string | string[] = Breakpoints.XSmall;
-
-  private readonly _disabled$ = new BehaviorSubject<boolean>(false);
-  public readonly disabled$: Observable<boolean> =
-    this._disabled$.asObservable();
-
-  @Input()
-  public set disabled(disabled: boolean) {
-    this._disabled$.next(disabled);
-  }
-
-  private readonly _expanded$ = new BehaviorSubject<boolean>(false);
-  public readonly expanded$: Observable<boolean> =
-    this._expanded$.asObservable();
-
-  @Input()
-  public set expanded(expanded: boolean) {
-    this._expanded$.next(expanded);
-  }
-
-  @Input()
-  public navListMatchOptions?: IsActiveMatchOptions | { exact: boolean };
-
-  public constructor(
-    public readonly router: Router,
-    public readonly breakpointObserver: BreakpointObserver,
-  ) {
-    super();
-  }
-
-  public ngOnInit(): void {
-    if (!(this.item.route || this.item.href || this.item.children)) {
-      this._disabled$.next(true);
-    } else {
-      this.addSubscription(
-        this.router.events.pipe(filter((e) => e instanceof NavigationEnd)),
-        (event: NavigationEnd) => {
-          this.item.navigationFinished?.(true);
+  public constructor() {
+    effect(() => {
+      const item = this.item();
+      if (!(item.route || item.href || item.children)) {
+        this.disabled.set(true);
+      } else {
+        this.router.events.pipe(filter((e): e is NavigationEnd => e instanceof NavigationEnd)).subscribe(event => {
+          item.navigationFinished?.(true);
           const url = event.urlAfterRedirects;
-          if (url && this.item?.children?.length) {
-            this._expanded$.next(
-              this.item.children.some(
-                (item: NavItem) => url.indexOf(`/${item.route}`) === 0,
-              ),
-            );
+          if (url && item?.children?.length) {
+            this.expanded.set(item.children.some((item: NavItem) => url.indexOf(`/${item.route}`) === 0));
           }
-        },
-      );
-    }
+        });
+      }
+    });
   }
 
   public onClick(event: MouseEvent): void {
-    if (this.item?.children?.length) {
+    if (this.item()?.children?.length) {
       event.preventDefault();
-      this._expanded$.next(!this._expanded$.value);
+      this.expanded.set(!this.expanded());
     }
   }
 }
