@@ -1,11 +1,4 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  EventEmitter,
-  Input,
-  Output,
-  ViewEncapsulation,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, computed, inject, input, output, ViewEncapsulation } from '@angular/core';
 import { TableCell } from '../../models/table-cell.model';
 import {
   CdkDrag,
@@ -16,8 +9,6 @@ import {
   CdkDropList,
   moveItemInArray,
 } from '@angular/cdk/drag-drop';
-import { Observable } from 'rxjs';
-import { map, take, withLatestFrom } from 'rxjs/operators';
 import { Lab900TableService } from '../../services/table.service';
 import { readPropValue } from '../../../utils/utils';
 import { MatMenuModule } from '@angular/material/menu';
@@ -49,32 +40,20 @@ import { MatButtonModule } from '@angular/material/button';
   ],
 })
 export class Lab900TableFilterMenuComponent {
-  public readonly tableCells$: Observable<TableCell[]>;
-  public readonly visibleCells$: Observable<TableCell[]>;
-  public readonly hiddenCells$: Observable<TableCell[]>;
+  private readonly tableService = inject(Lab900TableService);
+  /**
+   * Filterable table cells. If the cell is always visible, it will not be shown in the filter menu.
+   */
+  protected readonly filterableTableCells = computed(() =>
+    this.tableService.columns().filter((cell: TableCell) => !cell.alwaysVisible)
+  );
+  protected readonly visibleCells = computed(() => this.filterableTableCells().filter((cell: TableCell) => !cell.hide));
+  protected readonly hiddenCells = computed(() => this.filterableTableCells().filter((cell: TableCell) => !!cell.hide));
 
-  @Input()
-  public filterIcon = 'filter_alt';
+  public filterIcon = input<string>('filter_alt');
+  public toggleAndMoveColumns = input<boolean>(false);
 
-  @Input()
-  public toggleAndMoveColumns = false;
-
-  @Output()
-  public filterChanged: EventEmitter<TableCell[]> = new EventEmitter<
-    TableCell[]
-  >();
-
-  public constructor(private tableService: Lab900TableService) {
-    this.tableCells$ = this.tableService.columns$.pipe(
-      map((cells) => cells?.filter((cell: TableCell) => !cell.alwaysVisible)),
-    );
-    this.visibleCells$ = this.tableCells$.pipe(
-      map((cells) => cells?.filter((cell: TableCell) => !cell.hide)),
-    );
-    this.hiddenCells$ = this.tableCells$.pipe(
-      map((cells) => cells?.filter((cell: TableCell) => !!cell.hide)),
-    );
-  }
+  protected readonly filterChanged = output<TableCell[]>();
 
   public getCellLabel(cell: TableCell): string {
     return readPropValue(cell.label, cell);
@@ -83,27 +62,20 @@ export class Lab900TableFilterMenuComponent {
   public handleCheckboxClick(event: MouseEvent, cell: TableCell): void {
     event.stopPropagation();
     event.preventDefault();
-    this.tableService.columns$.pipe(take(1)).subscribe((tableCells) => {
-      const index = tableCells.findIndex((c) => c.key === cell.key);
-      tableCells[index].hide = !tableCells[index].hide;
-      this.filterChanged.emit(tableCells);
-    });
+    const tableCells = this.filterableTableCells();
+    const index = tableCells.findIndex(c => c.key === cell.key);
+    tableCells[index].hide = !tableCells[index].hide;
+    this.filterChanged.emit(tableCells);
   }
 
   public drop($event: CdkDragDrop<TableCell[]>): void {
-    this.visibleCells$
-      .pipe(withLatestFrom(this.tableService.columns$), take(1))
-      .subscribe(([visibleColumns, tableCells]) => {
-        moveItemInArray(
-          visibleColumns,
-          $event.previousIndex,
-          $event.currentIndex,
-        );
-        visibleColumns.forEach((cell, newColumnOrder) => {
-          const index = tableCells.findIndex((c) => c.key === cell.key);
-          tableCells[index].columnOrder = newColumnOrder;
-        });
-        this.filterChanged.emit(tableCells);
-      });
+    const tableCells = this.filterableTableCells();
+    const visibleColumns = this.visibleCells();
+    moveItemInArray(visibleColumns, $event.previousIndex, $event.currentIndex);
+    visibleColumns.forEach((cell, newColumnOrder) => {
+      const index = tableCells.findIndex(c => c.key === cell.key);
+      tableCells[index].columnOrder = newColumnOrder;
+    });
+    this.filterChanged.emit(tableCells);
   }
 }
