@@ -1,15 +1,19 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  ContentChild,
-  EventEmitter,
+  computed,
+  contentChild,
+  effect,
   inject,
+  input,
   Input,
-  OnDestroy,
-  Output,
+  model,
+  output,
+  signal,
   TemplateRef,
   TrackByFunction,
-  ViewChild,
+  untracked,
+  viewChild,
   ViewEncapsulation,
 } from '@angular/core';
 import { Lab900TableEmptyDirective } from '../../directives/table-empty.directive';
@@ -20,28 +24,13 @@ import { Lab900TableHeaderContentDirective } from '../../directives/table-header
 import { ActionButton } from '../../../button/models/action-button.model';
 import { Lab900TableTopContentDirective } from '../../directives/table-top-content.directive';
 import { MatTable, MatTableModule } from '@angular/material/table';
-import {
-  CdkDrag,
-  CdkDragDrop,
-  CdkDragHandle,
-  CdkDragPlaceholder,
-  CdkDropList,
-} from '@angular/cdk/drag-drop';
+import { CdkDrag, CdkDragDrop, CdkDragHandle, CdkDragPlaceholder, CdkDropList } from '@angular/cdk/drag-drop';
 import { ThemePalette } from '@angular/material/core';
 import { Lab900Sort } from '../../models/table-sort.model';
 import { Lab900TableTab } from '../../models/table-tabs.model';
-import { Lab900TableLeftFooterDirective } from '../../directives/table-left-footer.directive';
-import {
-  BehaviorSubject,
-  combineLatest,
-  Observable,
-  ReplaySubject,
-  Subscription,
-} from 'rxjs';
-import { filter, map, shareReplay, take, withLatestFrom } from 'rxjs/operators';
 import { Lab900TableService } from '../../services/table.service';
 import { Lab900TableHeaderComponent } from '../table-header/lab900-table-header.component';
-import { AsyncPipe, NgClass, NgTemplateOutlet } from '@angular/common';
+import { NgClass, NgTemplateOutlet } from '@angular/common';
 import { Lab900TableTabsComponent } from '../table-tabs/table-tabs.component';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { TableCellSelectComponent } from '../table-cell-select/table-cell-select.component';
@@ -70,6 +59,7 @@ export interface SelectableRows<T = any> {
   selectedItems?: T[];
   singleSelect?: boolean;
   hideSelectableRow?: (row: T) => boolean;
+  compareFn?: (o1: T, o2: T) => boolean;
 }
 
 @Component({
@@ -82,7 +72,6 @@ export interface SelectableRows<T = any> {
   standalone: true,
   imports: [
     Lab900TableHeaderComponent,
-    AsyncPipe,
     NgTemplateOutlet,
     Lab900TableTabsComponent,
     MatProgressBarModule,
@@ -99,34 +88,21 @@ export interface SelectableRows<T = any> {
     Lab900ActionButtonComponent,
   ],
 })
-export class Lab900TableComponent<T extends object = object, TabId = string>
-  implements OnDestroy
-{
-  private readonly tableService: Lab900TableService<T, TabId> =
-    inject(Lab900TableService);
-  private readonly footerSub: Subscription;
-
-  private readonly _fixedWidth$ = new BehaviorSubject<boolean>(false);
-  public readonly fixedWidth$ = combineLatest([
-    this.tableService.visibleColumns$,
-    this._fixedWidth$.asObservable(),
-  ]).pipe(
-    map(([columns, fixedWidth]) => {
-      fixedWidth || columns?.some((c) => !!c?.width);
-    }),
-  );
+export class Lab900TableComponent<T extends object = object, TabId = string> {
+  private readonly tableService = inject(Lab900TableService<T, TabId>);
 
   /**
    * This will respect the defined widths of the table cells
    * If any cell has a width defined, the table will have a fixed layout
    */
-  @Input()
-  public set fixedWidths(fixedWidth: boolean) {
-    this._fixedWidth$.next(fixedWidth);
-  }
+  public readonly _fixedWidths = input<boolean>(false, {
+    alias: 'fixedWidth',
+  });
+  protected readonly fixedWidths = computed(() => {
+    return this._fixedWidths() || this.tableService.visibleColumns().some(c => !!c?.width);
+  });
 
-  @ViewChild(MatTable)
-  public table?: MatTable<T>;
+  public readonly table = viewChild(MatTable<T>);
 
   @Input()
   public set tableCells(columns: TableCell<T>[]) {
@@ -143,263 +119,182 @@ export class Lab900TableComponent<T extends object = object, TabId = string>
     this.tableService.updateTabId(tabId);
   }
 
-  public get draggableRows(): boolean {
-    return (
-      this.tableActionsBack?.some((a) => !!a?.draggable) ||
-      this.tableActionsFront?.some((a) => !!a?.draggable)
-    );
-  }
-
-  private readonly _data$ = new ReplaySubject<T[] | null>();
-  public readonly data$: Observable<T[] | null>;
-
-  @Input()
-  public set data(value: T[] | null) {
-    this._data$.next(value);
-  }
-
-  @Input()
-  public tableClass?: string;
-
-  @Input()
-  public rowClass?: propFunction<T> | string;
-
-  @Input()
-  public rowColor?: propFunction<T> | string;
-
-  @Input()
-  public loading = false;
+  public readonly tableClass = input<string>('');
+  public readonly rowClass = input<propFunction<T> | string | undefined>(undefined);
+  public readonly rowColor = input<propFunction<T> | string | undefined>(undefined);
+  public readonly loading = model<boolean>(false);
 
   /**
    * Show a set of actions at the top of the table
    */
-  @Input()
-  public tableHeaderActions?: ActionButton<T>[];
+  public readonly tableHeaderActions = input<ActionButton<T>[]>([]);
 
   /**
    * Show a set of actions at the bottom of the table
    */
-  @Input()
-  public tableFooterActions?: ActionButton<T>[];
+  public readonly tableFooterActions = input<ActionButton<T>[]>([]);
 
   /**
    * Show a set of actions at the start of each row
    */
-  @Input()
-  public tableActionsFront?: TableRowAction<T>[];
+  public readonly tableActionsFront = input<TableRowAction<T>[]>([]);
 
   /**
    * Show a set of actions at the end of each row
    */
-  @Input()
-  public tableActionsBack?: TableRowAction<T>[];
+  public readonly tableActionsBack = input<TableRowAction<T>[]>([]);
 
-  @Input()
-  public selection?: SelectionModel<T>;
+  public readonly draggableRows = computed(() => {
+    return this.tableActionsBack()?.some(a => !!a?.draggable) || this.tableActionsFront()?.some(a => !!a?.draggable);
+  });
 
-  private readonly _selectableRows$ =
-    new BehaviorSubject<SelectableRows<T> | null>(null);
+  public readonly selectableRows = input<SelectableRows<T> | undefined>(undefined);
 
-  public readonly selectableRows$: Observable<SelectableRows<T>> =
-    this._selectableRows$
-      .asObservable()
-      .pipe(filter((selectableRows) => !!selectableRows));
-
-  @Input()
-  public set selectableRows(value: SelectableRows<T> | undefined) {
-    if (value && value.enabled) {
-      this._selectableRows$.next(value);
-      this.selection = new SelectionModel<any>(
-        !value?.singleSelect,
-        value?.selectedItems ?? [],
-      );
-    } else {
-      this._selectableRows$.next(null);
-    }
-  }
+  public readonly selection = signal<SelectionModel<T> | undefined>(undefined);
 
   /**
    * Show columns filter to hide/show columns
    */
-  @Input()
-  public toggleColumns = false;
+  public readonly toggleColumns = input<boolean>(false);
 
   /**
    * Show columns filter to hide/show columns AND show rearrange option
    * This overrides toggleColumns field
    */
-  @Input()
-  public toggleAndMoveColumns = false;
+  public readonly toggleAndMoveColumns = input<boolean>(false);
+  public readonly filterIcon = input<string>('filter_alt');
+  public readonly neverHideTable = input<boolean>(false);
+  public readonly disabled = model<boolean>(false);
 
   @Input()
-  public filterIcon = 'filter_alt';
-
-  @Input()
-  public neverHideTable = false;
-
-  @Input()
-  public disabled = false;
-
-  @Input()
-  public set sort(value: Lab900Sort[] | null) {
+  public set sort(value: Lab900Sort[] | undefined) {
     this.tableService.updateSorting(value);
   }
 
-  @Input()
-  public multiSort = false;
-
-  @Input()
-  public disableSort = false;
-
-  @Output()
-  public readonly sortChange = new EventEmitter<Lab900Sort[]>();
+  public readonly multiSort = input<boolean>(false);
+  public readonly disableSort = input<boolean>(false);
 
   /**
    * set max width of all columns, can be individually overwritten per tableCell
    * this will create a tooltip with the content of the cell, when the text overflows
    * @example: '300px'
    */
-  @Input()
-  public maxColumnWidth?: string;
+  public readonly maxColumnWidth = input<string | undefined>(undefined);
 
-  @Input()
-  public onRowClick?: (value: T, index: number, event: Event) => void;
+  public readonly onRowClick = input<((value: T, index: number, event: Event) => void) | undefined>(undefined);
 
-  @Input()
-  public preFooterTitle?: string;
-
-  @Input()
-  public stickyHeader?: boolean;
+  public readonly preFooterTitle = input<string | undefined>(undefined);
+  public readonly stickyHeader = input<boolean>(false);
+  public readonly trackByTableFn = input.required<TrackByFunction<T>>();
 
   @Input()
   public set disabledEditing(value: boolean) {
     this.tableService.updateDisableEditing(value);
   }
 
-  @Output()
-  public activeTabIdChange = new EventEmitter<TabId>();
+  // outputs
+  public readonly activeTabIdChange = output<TabId>();
+  public readonly selectionChanged = output<SelectionModel<T>>();
+  public readonly rowSelectToggle = output<T>();
+  public readonly tableCellsFiltered = output<TableCell<T>[]>();
+  public readonly tableRowOrderChange = output<CdkDragDrop<T[]>>();
+  public readonly cellValueChanged = output<CellValueChangeEvent<T>>();
+  public readonly sortChange = output<Lab900Sort[]>();
 
-  @Output()
-  public readonly selectionChanged = new EventEmitter<SelectionModel<T>>();
+  // content children
+  protected readonly emptyTableTemplate = contentChild(Lab900TableEmptyDirective, {
+    read: TemplateRef,
+  });
+  protected readonly disabledTableTemplate = contentChild(Lab900TableDisabledDirective, {
+    read: TemplateRef,
+  });
+  protected readonly tableHeaderContent = contentChild(Lab900TableHeaderContentDirective, {
+    read: TemplateRef,
+  });
+  protected readonly tableTopContent = contentChild(Lab900TableTopContentDirective, {
+    read: TemplateRef,
+  });
 
-  @Output()
-  public readonly rowSelectToggle = new EventEmitter<T>();
+  public readonly visibleColumns = this.tableService.visibleColumns;
+  public readonly displayedColumns = computed(() => this.getDisplayedColumns());
+  public readonly tabId = this.tableService.tabId;
+  public readonly tabs = this.tableService.tabs;
+  public readonly showCellFooters = computed(() => this.visibleColumns().some(c => !!c?.footer));
 
-  @Output()
-  public readonly tableCellsFiltered = new EventEmitter<TableCell<T>[]>();
-
-  @Output()
-  public readonly tableRowOrderChange = new EventEmitter<CdkDragDrop<T[]>>();
-
-  @Output()
-  public readonly cellValueChanged = new EventEmitter<
-    CellValueChangeEvent<T>
-  >();
-
-  @ContentChild(Lab900TableEmptyDirective, { read: TemplateRef })
-  public emptyTableTemplate?: TemplateRef<any>;
-
-  @ContentChild(Lab900TableDisabledDirective, { read: TemplateRef })
-  public disabledTableTemplate?: TemplateRef<any>;
-
-  @ContentChild(Lab900TableHeaderContentDirective, { read: TemplateRef })
-  public tableHeaderContent?: TemplateRef<any>;
-
-  @ContentChild(Lab900TableTopContentDirective, { read: TemplateRef })
-  public tableTopContent?: TemplateRef<any>;
-
-  @ContentChild(Lab900TableLeftFooterDirective, { read: TemplateRef })
-  public footerLeftContent?: TemplateRef<any>;
-
-  @Input({ required: true })
-  public trackByTableFn!: TrackByFunction<T>;
-
-  public readonly showCellFooters$: Observable<boolean>;
-
-  public readonly visibleColumns$: Observable<TableCell<T>[]> =
-    this.tableService.visibleColumns$;
-  public readonly tabId$: Observable<TabId> = this.tableService.tabId$;
-  public readonly tabs$: Observable<Lab900TableTab<TabId, T>[]> =
-    this.tableService.tabs$;
-  public readonly displayedColumns$: Observable<string[]>;
+  public readonly data = model<T[] | null | undefined>(undefined);
+  public readonly publicData = computed(() => {
+    const hideSelectableRow = this.selectableRows()?.hideSelectableRow;
+    const data = this.data();
+    if (hideSelectableRow) {
+      data?.map(v => ({
+        ...v,
+        _hideSelectableRow: hideSelectableRow(v),
+      }));
+    }
+    return data;
+  });
 
   public constructor() {
-    this.showCellFooters$ = this.visibleColumns$.pipe(
-      map((columns) => !!columns?.some((c) => !!c?.footer)),
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
-    this.displayedColumns$ = combineLatest([
-      this.visibleColumns$,
-      this._selectableRows$.asObservable(),
-    ]).pipe(
-      map(([columns, selectableRows]) =>
-        this.getDisplayedColumns(columns, selectableRows),
-      ),
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
-    this.data$ = combineLatest([
-      this._selectableRows$.asObservable(),
-      this._data$.asObservable(),
-    ]).pipe(
-      map(([options, data]) =>
-        options?.hideSelectableRow
-          ? data?.map((v) => ({
-              ...v,
-              _hideSelectableRow: options.hideSelectableRow(v),
-            }))
-          : data,
-      ),
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
-
-    /**
-     * Fix for rendering async footers
-     */
-    this.footerSub = this.data$
-      .pipe(withLatestFrom(this.showCellFooters$))
-      .subscribe(([data, showFooter]) => {
-        if (this.table) {
-          this.table.removeFooterRowDef(null);
-          if (data?.length && showFooter) {
-            this.table.renderRows();
-          }
+    effect(() => {
+      const table = untracked(this.table);
+      if (table) {
+        table.removeFooterRowDef(null as unknown as any);
+        if (this.data()?.length && this.showCellFooters()) {
+          table.renderRows();
         }
-      });
-  }
+      }
+    });
 
-  public ngOnDestroy(): void {
-    this.footerSub.unsubscribe();
+    effect(
+      () => {
+        const selectableRows = this.selectableRows();
+        if (selectableRows?.enabled && !untracked(this.selection)) {
+          this.selection.set(
+            new SelectionModel(
+              !selectableRows.singleSelect,
+              selectableRows.selectedItems,
+              true,
+              selectableRows.compareFn
+            )
+          );
+        }
+      },
+      { allowSignalWrites: true }
+    );
   }
 
   public handleSelectAll(checked: boolean): void {
-    this.selection.clear();
-    if (checked) {
-      this.data$.pipe(take(1)).subscribe((data) => {
+    const selection = this.selection();
+    if (selection) {
+      selection.clear();
+      const data = this.data();
+      if (checked) {
         if (data?.length) {
-          this.selection.select(
-            ...data.filter((row) => !(row as any)._hideSelectableRow),
-          );
-          this.selectionChanged.emit(this.selection);
+          selection.select(...data.filter(row => !(row as any)?._hideSelectableRow));
+          this.selectionChanged.emit(selection);
         }
-      });
-    } else {
-      this.selectionChanged.emit(this.selection);
+      } else {
+        this.selectionChanged.emit(selection);
+      }
     }
   }
 
   public handleSelectRow(row: T): void {
-    this.selection.toggle(row);
-    this.selectionChanged.emit(this.selection);
-    this.rowSelectToggle.emit(row);
+    const selection = this.selection();
+    if (selection) {
+      selection.toggle(row);
+      this.selectionChanged.emit(selection);
+      this.rowSelectToggle.emit(row);
+    }
   }
 
   public isRowSelected(row: T): boolean {
-    return this.selection?.isSelected(row);
+    return !!this.selection()?.isSelected(row);
   }
 
   public getRowClasses(row: T, index: number): string {
     const classes: string[] = [];
-    if (typeof this.onRowClick === 'function') {
+    if (typeof this.onRowClick() === 'function') {
       classes.push('lab900-row-clickable');
     }
     if (index % 2 === 0) {
@@ -407,37 +302,28 @@ export class Lab900TableComponent<T extends object = object, TabId = string>
     } else {
       classes.push('lab900-row-odd');
     }
-    classes.push(
-      (typeof this.rowClass === 'function'
-        ? this.rowClass(row)
-        : this.rowClass) ?? '',
-    );
+    const rowClass = this.rowClass();
+    if (rowClass) {
+      classes.push((typeof rowClass === 'function' ? rowClass(row) : rowClass) ?? '');
+    }
     return classes.join(' ') || '';
   }
 
   public getRowColor(row: T): string {
-    return (
-      (typeof this.rowColor === 'function'
-        ? this.rowColor(row)
-        : this.rowColor) ?? ''
-    );
+    const rowColor = this.rowColor();
+    return (typeof rowColor === 'function' ? rowColor(row) : rowColor) ?? '';
   }
 
   public handleRowClick(event: Event, row: T, index: number): void {
-    if (typeof this.onRowClick === 'function') {
-      this.onRowClick(row, index, event);
+    const rowClick = this.onRowClick();
+    if (typeof rowClick === 'function') {
+      rowClick(row, index, event);
     }
   }
 
-  public trackCellFn(_, item: TableCell<T>): string {
-    return item.key;
-  }
-
   public handleHeaderClick(cell: TableCell<T>): void {
-    if (!this.disableSort && cell.sortable) {
-      this.tableService.updateColumnSorting(cell, this.multiSort, (sort) =>
-        this.sortChange.emit(sort),
-      );
+    if (!this.disableSort() && cell.sortable) {
+      this.tableService.updateColumnSorting(cell, this.multiSort(), sort => this.sortChange.emit(sort ?? []));
     }
   }
 
@@ -451,19 +337,16 @@ export class Lab900TableComponent<T extends object = object, TabId = string>
     this.activeTabIdChange.emit(id);
   }
 
-  private getDisplayedColumns(
-    columns: TableCell<T>[],
-    selectableRows: SelectableRows<T>,
-  ): string[] {
-    const displayColumns = columns?.map((c) => c.key);
-    if (this.tableActionsFront?.length) {
+  private getDisplayedColumns(): string[] {
+    const displayColumns = this.visibleColumns()?.map(c => c.key);
+    if (this.tableActionsFront().length) {
       displayColumns.unshift('actions-front');
     }
-    if (this.tableActionsBack?.length) {
+    if (this.tableActionsBack().length) {
       displayColumns.push('actions-back');
     }
-    if (selectableRows) {
-      if (selectableRows?.position === 'right') {
+    if (this.selectableRows()?.enabled) {
+      if (this.selectableRows()?.position === 'right') {
         displayColumns.push('select');
       } else {
         displayColumns.unshift('select');
