@@ -1,19 +1,6 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  ViewEncapsulation,
-} from '@angular/core';
-import {
-  BehaviorSubject,
-  combineLatest,
-  Observable,
-  ReplaySubject,
-} from 'rxjs';
-import { map, shareReplay } from 'rxjs/operators';
+import { ChangeDetectionStrategy, Component, computed, input, ViewEncapsulation } from '@angular/core';
 import { BreadCrumb } from '../../models/bread-crumb.model';
 import { readPropValue } from '../../../utils/utils';
-import { AsyncPipe } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterLink } from '@angular/router';
 
@@ -22,43 +9,14 @@ import { RouterLink } from '@angular/router';
   standalone: true,
   templateUrl: './bread-crumb-item.component.html',
   styleUrls: ['./bread-crumb-item.component.scss'],
-  imports: [TranslateModule, AsyncPipe, RouterLink],
+  imports: [TranslateModule, RouterLink],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
 })
 export class BreadCrumbItemComponent {
-  private readonly _item$ = new ReplaySubject<BreadCrumb>();
-  private readonly item$: Observable<BreadCrumb> = this._item$
-    .asObservable()
-    .pipe(shareReplay({ bufferSize: 1, refCount: true }));
-
-  @Input({ required: true })
-  public set item(item: BreadCrumb) {
-    this._item$.next(item);
-  }
-
-  private readonly _data$ = new BehaviorSubject<any>(undefined);
-  private readonly data$ = this._data$
-    .asObservable()
-    .pipe(shareReplay({ bufferSize: 1, refCount: true }));
-
-  @Input()
-  public set data(data: any) {
-    this._data$.next(data);
-  }
-
-  public readonly title$: Observable<string>;
-  public readonly route$: Observable<string>;
-  public readonly queryParams$: Observable<object>;
-
-  public constructor() {
-    const stream = combineLatest([this.item$, this.data$]).pipe(
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
-    this.title$ = stream.pipe(map(([a, d]) => readPropValue(a.title, d)));
-    this.route$ = stream.pipe(map(([a, d]) => readPropValue(a.route, d)));
-    this.queryParams$ = stream.pipe(
-      map(([a, d]) => readPropValue(a.queryParams, d)),
-    );
-  }
+  public readonly item = input.required<BreadCrumb>();
+  public readonly data = input<any>(undefined);
+  protected readonly title = computed(() => readPropValue(this.item().title, this.data()));
+  protected readonly route = computed(() => readPropValue(this.item().route, this.data()));
+  protected readonly queryParams = computed(() => readPropValue(this.item().queryParams, this.data()));
 }
