@@ -7,27 +7,20 @@ import { MatIcon } from '@angular/material/icon';
 import { AsyncPipe } from '@angular/common';
 import { TranslateModule } from '@ngx-translate/core';
 import { PreventDoubleClickDirective } from '../../directives/preventDoubleClick.directive';
+import { map } from 'rxjs/operators';
 
 @Component({
   selector: 'lab900-action-button-menu',
   templateUrl: './lab900-action-button-menu.component.html',
   standalone: true,
-  imports: [
-    MatMenu,
-    MatIcon,
-    MatMenuItem,
-    MatMenuTrigger,
-    AsyncPipe,
-    TranslateModule,
-    PreventDoubleClickDirective,
-  ],
+  imports: [MatMenu, MatIcon, MatMenuItem, MatMenuTrigger, AsyncPipe, TranslateModule, PreventDoubleClickDirective],
 })
 export class Lab900ActionButtonMenuComponent {
   @ViewChild('actionMenu', { static: true })
-  public actionMenu: MatMenu;
+  public actionMenu?: MatMenu;
 
   @Input({ required: true })
-  public actions: ActionButton[];
+  public actions!: ActionButton[];
 
   @Input()
   public data: any;
@@ -37,7 +30,7 @@ export class Lab900ActionButtonMenuComponent {
   }
 
   public getDisabled(action: ActionButton): Observable<boolean> {
-    return coerceObservable(readPropValue(action.disabled, this.data));
+    return coerceObservable(readPropValue(action.disabled, this.data)).pipe(map(v => !!v));
   }
 
   public doAction(e: Event, action: ActionButton): void {
@@ -48,10 +41,14 @@ export class Lab900ActionButtonMenuComponent {
   }
 
   public getPrefixIcon(action: ActionButton): string {
-    return readPropValue(action.prefixIcon, this.data);
+    return readPropValue(action.prefixIcon, this.data) ?? '';
   }
 
   public getSuffixIcon(action: ActionButton): string {
-    return readPropValue(action.suffixIcon, this.data);
+    return readPropValue(action.suffixIcon, this.data) ?? '';
+  }
+
+  public getSubActionHidden(subAction: ActionButton): Observable<boolean> {
+    return coerceObservable(readPropValue(subAction.hide, this.data)).pipe(map(v => !!v));
   }
 }
