@@ -1,18 +1,10 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import {
-  ActionButton,
-  ActionButtonComponent,
-} from '../../models/action-button.model';
+import { ActionButton, ActionButtonComponent } from '../../models/action-button.model';
 import { MatTooltip, TooltipPosition } from '@angular/material/tooltip';
 import { ThemePalette } from '@angular/material/core';
 import { Lab900ButtonType } from '../../models/button.model';
 import { coerceObservable, readPropValue } from '../../../utils/utils';
-import {
-  BehaviorSubject,
-  combineLatest,
-  Observable,
-  ReplaySubject,
-} from 'rxjs';
+import { BehaviorSubject, combineLatest, Observable, ReplaySubject } from 'rxjs';
 import { filter, map, shareReplay, switchMap, take } from 'rxjs/operators';
 import { AsyncPipe } from '@angular/common';
 import { Lab900ButtonComponent } from '../button/button.component';
@@ -35,15 +27,14 @@ import { TranslateModule } from '@ngx-translate/core';
     TranslateModule,
   ],
 })
-export class Lab900ActionButtonComponent<T = any>
-  implements ActionButtonComponent<T>
-{
+export class Lab900ActionButtonComponent<T = any> implements ActionButtonComponent<T> {
   private readonly _action$ = new ReplaySubject<ActionButton<T>>();
   public readonly action$: Observable<ActionButton> = this._action$
     .asObservable()
     .pipe(shareReplay({ bufferSize: 1, refCount: true }));
-  public readonly tooltipPosition$: Observable<TooltipPosition> =
-    this.action$.pipe(map((action) => action.tooltip?.position ?? 'above'));
+  public readonly tooltipPosition$: Observable<TooltipPosition> = this.action$.pipe(
+    map(action => action.tooltip?.position ?? 'above')
+  );
 
   @Input({ required: true })
   public set action(action: ActionButton<T>) {
@@ -51,9 +42,7 @@ export class Lab900ActionButtonComponent<T = any>
   }
 
   private readonly _data$ = new BehaviorSubject<any>(undefined);
-  public readonly data$ = this._data$
-    .asObservable()
-    .pipe(shareReplay({ bufferSize: 1, refCount: true }));
+  public readonly data$ = this._data$.asObservable().pipe(shareReplay({ bufferSize: 1, refCount: true }));
 
   @Input()
   public set data(data: any) {
@@ -63,43 +52,36 @@ export class Lab900ActionButtonComponent<T = any>
   @Output()
   public valueChanged = new EventEmitter<any>();
 
-  public readonly buttonType$: Observable<Lab900ButtonType | 'toggle'>;
+  public readonly buttonType$: Observable<Lab900ButtonType | 'toggle' | undefined>;
   public readonly color$: Observable<ThemePalette>;
   public readonly label$: Observable<string>;
   public readonly hidden$: Observable<boolean>;
 
   private readonly _disabled$ = new BehaviorSubject<boolean>(false);
   public readonly disabled$: Observable<boolean>;
-  public readonly suffixIcon$: Observable<string>;
-  public readonly prefixIcon$: Observable<string>;
-  public readonly containerClass$: Observable<string>;
+  public readonly suffixIcon$: Observable<string | undefined>;
+  public readonly prefixIcon$: Observable<string | undefined>;
+  public readonly containerClass$: Observable<string | undefined>;
 
   public constructor() {
-    const stream = combineLatest([this.action$, this.data$]).pipe(
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
+    const stream = combineLatest([this.action$, this.data$]).pipe(shareReplay({ bufferSize: 1, refCount: true }));
 
     this.buttonType$ = stream.pipe(map(([a, d]) => readPropValue(a.type, d)));
     this.color$ = stream.pipe(map(([a, d]) => readPropValue(a.color, d)));
     this.label$ = stream.pipe(map(([a, d]) => readPropValue(a.label, d)));
-    this.hidden$ = stream.pipe(
-      switchMap(([a, d]) => coerceObservable(readPropValue(a.hide, d))),
-    );
+    this.hidden$ = stream
+      .pipe(switchMap(([a, d]) => coerceObservable(readPropValue(a.hide, d))))
+      .pipe(map(hidden => !!hidden));
     this.disabled$ = combineLatest([stream, this._disabled$]).pipe(
       switchMap(([[a, d], disabled]) => {
         return coerceObservable(disabled || readPropValue(a.disabled, d));
       }),
+      map(disabled => !!disabled)
     );
 
-    this.prefixIcon$ = stream.pipe(
-      map(([a, d]) => readPropValue(a.prefixIcon, d)),
-    );
-    this.suffixIcon$ = stream.pipe(
-      map(([a, d]) => readPropValue(a.suffixIcon, d)),
-    );
-    this.containerClass$ = stream.pipe(
-      map(([a, d]) => readPropValue(a.containerClass, d)),
-    );
+    this.prefixIcon$ = stream.pipe(map(([a, d]) => readPropValue(a.prefixIcon, d)));
+    this.suffixIcon$ = stream.pipe(map(([a, d]) => readPropValue(a.suffixIcon, d)));
+    this.containerClass$ = stream.pipe(map(([a, d]) => readPropValue(a.containerClass, d)));
   }
 
   public doAction(e: Event): void {
@@ -107,10 +89,10 @@ export class Lab900ActionButtonComponent<T = any>
     combineLatest([this.action$, this.data$])
       .pipe(
         take(1),
-        filter(([a]) => !!a.action),
+        filter(([a]) => !!a.action)
       )
       .subscribe(([a, d]) => {
-        a.action(d, e, this);
+        a.action?.(d, e, this);
       });
   }
 
