@@ -1,7 +1,1 @@
-export type Lab900ButtonType =
-  | 'raised'
-  | 'stroked'
-  | 'flat'
-  | 'icon'
-  | 'fab'
-  | 'mini-fab';
+export type Lab900ButtonType = 'raised' | 'stroked' | 'flat' | 'icon' | 'fab' | 'mini-fab';
