@@ -1,11 +1,3 @@
-import { Injectable, Input } from '@angular/core';
-
-export interface CustomComponent<T> {
-  data: T;
-}
-
-@Injectable()
-export abstract class CustomComponentAbstract<T> implements CustomComponent<T> {
-  @Input()
-  public data: T;
+export interface MergerItemComponent<T = any> {
+  data?: T;
 }
