@@ -1,21 +1,14 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  ViewEncapsulation,
-} from '@angular/core';
+import { ChangeDetectionStrategy, Component, computed, input, ViewEncapsulation } from '@angular/core';
 import { NavItem, NavItemGroup } from '../../models/nav-item.model';
 import { IsActiveMatchOptions } from '@angular/router';
-import { Observable, ReplaySubject } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { AsyncPipe } from '@angular/common';
 import { IconComponent } from '../icon/icon.component';
 import { TranslateModule } from '@ngx-translate/core';
 import { MatListModule } from '@angular/material/list';
 import { NavItemComponent } from '../nav-item/nav-item.component';
+import { Breakpoints } from '@angular/cdk/layout';
 
 const hide = (i: { hide?: (() => boolean) | boolean }): boolean => {
-  return typeof i?.hide === 'function' ? i.hide() : i?.hide ?? false;
+  return typeof i?.hide === 'function' ? i.hide() : (i?.hide ?? false);
 };
 
 @Component({
@@ -25,55 +18,35 @@ const hide = (i: { hide?: (() => boolean) | boolean }): boolean => {
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
-  imports: [
-    AsyncPipe,
-    IconComponent,
-    TranslateModule,
-    MatListModule,
-    NavItemComponent,
-  ],
+  imports: [IconComponent, TranslateModule, MatListModule, NavItemComponent],
 })
 export class Lab900NavListComponent {
-  private readonly _navItemGroups$ = new ReplaySubject<NavItemGroup[]>();
-  public readonly filtersGroups$: Observable<NavItemGroup[]> =
-    this._navItemGroups$.asObservable().pipe(
-      map(([...groups]) => {
-        return groups
-          .filter((g) => !hide(g))
-          .map((g) => {
-            g.items = this.filterNavItems(g.items);
-            return g;
-          })
-          .filter((g) => !!g.items?.length);
-      }),
-    );
+  public readonly navItemGroups = input.required<NavItemGroup[]>();
+  public readonly indentLevels = input<boolean>(true);
+  public readonly showLevelArrows = input<boolean>(false);
+  public readonly allowOverlayMenuUntil = input<string | string[]>(Breakpoints.XSmall);
+  public readonly routeMatchOptions = input<IsActiveMatchOptions | { exact: boolean } | undefined>(undefined);
 
-  @Input({ required: true })
-  public set navItemGroups(value: NavItemGroup[]) {
-    this._navItemGroups$.next(value);
-  }
-
-  @Input()
-  public indentLevels = true;
-
-  @Input()
-  public showLevelArrows = false;
-
-  @Input()
-  public allowOverlayMenuUntil: string | string[] = 'xs';
-
-  @Input()
-  public routeMatchOptions?: IsActiveMatchOptions | { exact: boolean };
+  public readonly filteredNavItemGroups = computed(() => {
+    const groups = this.navItemGroups();
+    return groups
+      .filter(g => !hide(g))
+      .map(g => {
+        g.items = this.filterNavItems(g.items);
+        return g;
+      })
+      .filter(g => !!g.items?.length);
+  });
 
   private filterNavItems(items: NavItem[]): NavItem[] {
     return [...items]
-      .filter((i) => !hide(i))
-      .map((i) => {
+      .filter(i => !hide(i))
+      .map(i => {
         if (i?.children?.length) {
           i.children = this.filterNavItems(i.children);
         }
         return i;
       })
-      .filter((i) => !!i?.children?.length || i?.route || i?.href);
+      .filter(i => !!i?.children?.length || i?.route || i?.href);
   }
 }
