@@ -4,9 +4,7 @@ export interface NavItem {
   label: string;
   icon?: Icon;
   route?: string;
-  routeQueryParams?: {
-    [k: string]: any;
-  };
+  routeQueryParams?: Record<string, any>;
   href?: { url: string; target?: '_self' | '_blank' };
   children?: NavItem[];
   childrenInOverlay?: boolean;
