@@ -1,32 +1,23 @@
-import {
-  BehaviorSubject,
-  combineLatest,
-  Observable,
-  ReplaySubject,
-  Subscription,
-} from 'rxjs';
+import { Subscription } from 'rxjs';
 import { TableCell } from '../models/table-cell.model';
 import {
   AfterViewInit,
+  computed,
   Directive,
   ElementRef,
   inject,
-  Input,
+  input,
+  model,
   NgZone,
   OnDestroy,
-  ViewChild,
+  signal,
+  viewChild,
 } from '@angular/core';
-import { map, shareReplay } from 'rxjs/operators';
 import { readPropValue } from '../../utils/utils';
-import { TooltipPosition } from '@angular/material/tooltip';
 import { Lab900TableService } from '../services/table.service';
 
 @Directive()
-export abstract class CellRendererAbstract<
-    CellRenderOptions = any,
-    T = any,
-    V = any,
-  >
+export abstract class CellRendererAbstract<CellRenderOptions = any, T = any, V = any>
   implements AfterViewInit, OnDestroy
 {
   protected readonly elm: ElementRef<HTMLElement> = inject(ElementRef);
@@ -36,65 +27,35 @@ export abstract class CellRendererAbstract<
   private observer?: ResizeObserver;
   private observerSub?: Subscription;
 
-  protected readonly _columnConfig$ = new ReplaySubject<TableCell<T>>();
-  public readonly columnConfig$: Observable<TableCell<T, CellRenderOptions>> =
-    this._columnConfig$
-      .asObservable()
-      .pipe(shareReplay({ bufferSize: 1, refCount: true }));
+  public readonly columnConfig = input.required<TableCell<T, CellRenderOptions>>();
+  public readonly data = model.required<T>();
 
-  @Input({ required: true })
-  public set columnConfig(value: TableCell<T, CellRenderOptions>) {
-    this._columnConfig$.next(value);
-  }
+  public readonly renderOptions = computed<CellRenderOptions | undefined>(() => this.columnConfig().cellRenderOptions);
 
-  protected readonly _data$ = new ReplaySubject<T>();
-  public readonly data$ = this._data$.asObservable();
+  public readonly handleValueChanged = input<((value: V, cell: TableCell<T>, row: T) => void) | undefined>(undefined);
 
-  @Input({ required: true })
-  public set data(value: T) {
-    this._data$.next(value);
-  }
+  public readonly cellInnerElm = viewChild('.lab900-cell-value', {
+    read: ElementRef,
+  });
 
-  @Input()
-  public handleValueChanged?: (value: V, cell: TableCell<T>, row: T) => void;
+  public readonly cellValue = computed(() => this.getCellValue());
 
-  @ViewChild('.lab900-cell-value', { static: false, read: ElementRef })
-  public cellInnerElm?: ElementRef;
-
-  public readonly rendererOptions$: Observable<CellRenderOptions | undefined> =
-    this.columnConfig$.pipe(
-      map((config) => config.cellRenderOptions),
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
-
-  public readonly cellValue$: Observable<V> = this.getCellValue();
-
-  private readonly textOverflowing$ = new BehaviorSubject<boolean>(false);
-  public readonly tooltip$: Observable<string | undefined> = combineLatest([
-    this.columnConfig$,
-    this.data$,
-    this.cellValue$,
-    this.textOverflowing$,
-  ]).pipe(
-    map(([config, data, cellValue, textOverflowing]) => {
-      if (
-        config.cellTooltip?.text &&
-        (!config.cellTooltip?.onlyOnOverflow || textOverflowing)
-      ) {
-        return readPropValue(config.cellTooltip.text, data);
-      } else if (textOverflowing) {
-        return String(cellValue);
-      }
-      return undefined;
-    }),
+  private readonly textOverflowing = signal<boolean>(false);
+  public readonly tooltip = computed(() => {
+    const config = this.columnConfig();
+    const data = this.data();
+    const cellValue = this.cellValue();
+    const textOverflowing = this.textOverflowing();
+    if (config.cellTooltip?.text && (!config.cellTooltip?.onlyOnOverflow || textOverflowing)) {
+      return readPropValue(config.cellTooltip.text, data);
+    } else if (textOverflowing) {
+      return String(cellValue);
+    }
+    return undefined;
+  });
+  public readonly tooltipPosition = computed(
+    () => this.columnConfig().cellTooltip?.tooltipOptions?.tooltipPosition ?? 'below'
   );
-
-  public readonly tooltipPosition$: Observable<TooltipPosition> =
-    this.columnConfig$.pipe(
-      map(
-        (cell) => cell.cellTooltip?.tooltipOptions?.tooltipPosition ?? 'below',
-      ),
-    );
 
   public ngAfterViewInit(): void {
     this.observeCellContentOverflow();
@@ -105,35 +66,24 @@ export abstract class CellRendererAbstract<
     this.observer?.unobserve(this.elm.nativeElement);
   }
 
-  protected getCellValue(): Observable<V> {
-    const sources$ = {
-      config: this.columnConfig$,
-      data: this.data$,
-      disableEditing: this.tableService.disableEditing$,
-    };
+  protected getCellValue(): V | string {
+    const config = this.columnConfig();
+    const data = this.data();
+    const disableEditing = this.tableService.disableEditing();
 
-    return combineLatest(sources$).pipe(
-      map(({ config, data, disableEditing }) => {
-        const value = this.cellFormatter(config, data);
-        if (
-          !value &&
-          config.cellEditorOptions?.placeholder &&
-          !disableEditing &&
-          !config.cellEditorOptions?.disablePlaceholderOutsideEditor &&
-          !config.cellEditorOptions?.disabled?.(data)
-        ) {
-          this.elm.nativeElement
-            .querySelector('.lab900-cell-value')
-            ?.classList.add('value-is-placeholder');
-          return config.cellEditorOptions?.placeholder;
-        }
-        this.elm.nativeElement
-          .querySelector('.lab900-cell-value')
-          ?.classList.remove('value-is-placeholder');
-        return value;
-      }),
-      shareReplay({ bufferSize: 1, refCount: true }),
-    );
+    const value = this.cellFormatter(config, data);
+    if (
+      !value &&
+      config.cellEditorOptions?.placeholder &&
+      !disableEditing &&
+      !config.cellEditorOptions?.disablePlaceholderOutsideEditor &&
+      !config.cellEditorOptions?.disabled?.(data)
+    ) {
+      this.elm.nativeElement.querySelector('.lab900-cell-value')?.classList.add('value-is-placeholder');
+      return config.cellEditorOptions?.placeholder;
+    }
+    this.elm.nativeElement.querySelector('.lab900-cell-value')?.classList.remove('value-is-placeholder');
+    return value;
   }
 
   protected cellFormatter(cell: TableCell<T>, data: T): any {
@@ -157,12 +107,10 @@ export abstract class CellRendererAbstract<
   protected observeCellContentOverflow(): void {
     this.ngZone.runOutsideAngular(() => {
       this.observer?.unobserve(this.elm.nativeElement);
-      this.observer = new ResizeObserver((entries) => {
-        const innerScrollWidth =
-          this.elm.nativeElement.querySelector('.lab900-cell-value')
-            ?.scrollWidth ?? 0;
+      this.observer = new ResizeObserver(entries => {
+        const innerScrollWidth = this.elm.nativeElement.querySelector('.lab900-cell-value')?.scrollWidth ?? 0;
         const maxWidth = (entries[0].target as any).offsetWidth;
-        this.textOverflowing$.next(innerScrollWidth > maxWidth);
+        this.textOverflowing.set(innerScrollWidth > maxWidth);
       });
       this.observer.observe(this.elm.nativeElement);
     });
