@@ -1,13 +1,6 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  ViewEncapsulation,
-} from '@angular/core';
-import { AsyncPipe } from '@angular/common';
+import { ChangeDetectionStrategy, Component, computed, ViewEncapsulation } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
 import { CellWithClearingRendererOptions } from './cell-with-clearing-renderer.options';
-import { combineLatest } from 'rxjs';
-import { map, take, withLatestFrom } from 'rxjs/operators';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { TranslateModule } from '@ngx-translate/core';
 import { MatButtonModule } from '@angular/material/button';
@@ -16,13 +9,7 @@ import { CellRendererAbstract } from '../cell-renderer.abstract';
 @Component({
   selector: 'lab900-cell-with-clearing-renderer',
   standalone: true,
-  imports: [
-    AsyncPipe,
-    MatIconModule,
-    MatTooltipModule,
-    TranslateModule,
-    MatButtonModule,
-  ],
+  imports: [MatIconModule, MatTooltipModule, TranslateModule, MatButtonModule],
   templateUrl: './cell-with-clearing-renderer.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
@@ -41,38 +28,24 @@ import { CellRendererAbstract } from '../cell-renderer.abstract';
     `,
   ],
 })
-export class CellWithClearingRendererComponent<
-  T = any,
-> extends CellRendererAbstract<CellWithClearingRendererOptions, T> {
-  public readonly tooltipWithoutHtml$ = this.tooltip$.pipe(
-    map((tooltip) => {
-      return tooltip?.replace(/<[^>]*>/g, '');
-    }),
-  );
-  public readonly disabled$ = combineLatest([
-    this.columnConfig$,
-    this.data$,
-  ]).pipe(
-    map(([config, data]) => {
-      const value = this.cellFormatter(config, data);
-      const disabled = config?.cellEditorOptions?.disabled?.(data) ?? false;
-      return !value || disabled;
-    }),
-  );
+export class CellWithClearingRendererComponent<T = any> extends CellRendererAbstract<
+  CellWithClearingRendererOptions,
+  T
+> {
+  public readonly tooltipWithHtml = computed(() => this.tooltip()?.replace(/<[^>]*>/g, ''));
+  public readonly disabled = computed(() => {
+    const value = this.cellFormatter(this.columnConfig(), this.data());
+    const disabled = this.columnConfig()?.cellEditorOptions?.disabled?.(this.data()) ?? false;
+    return value || disabled;
+  });
 
   public handleClear(event: MouseEvent): void {
     event.preventDefault();
     event.stopImmediatePropagation();
-    this.columnConfig$
-      .pipe(take(1), withLatestFrom(this.data$))
-      .subscribe(([config, data]) => {
-        if (!this.handleValueChanged) {
-          throw Error(
-            `No handleValueChanged method provided for column ${config.key}`,
-          );
-        }
-        this.data = { ...data, [config.key]: null };
-        this.handleValueChanged?.(null, config, data);
-      });
+    const valueChanged = this.handleValueChanged();
+    if (!valueChanged) {
+      throw Error(`No handleValueChanged method provided for column ${this.columnConfig().key}`);
+    }
+    valueChanged(null, this.columnConfig(), this.data());
   }
 }
