@@ -1,54 +1,35 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  ViewEncapsulation,
-} from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component, computed, ViewEncapsulation } from '@angular/core';
 import { CellRendererAbstract } from '../cell-renderer.abstract';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { CheckboxCellRendererOptions } from './checkbox-cell-renderer.options';
-import { combineLatest } from 'rxjs';
-import { map, take, withLatestFrom } from 'rxjs/operators';
 import { MatTooltipModule } from '@angular/material/tooltip';
 
 @Component({
   selector: 'lab900-checkbox-cell-renderer',
   standalone: true,
-  imports: [CommonModule, MatCheckboxModule, MatTooltipModule],
+  imports: [MatCheckboxModule, MatTooltipModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
-  template: `@if (rendererOptions$ | async; as options) {
+  template: `
     <mat-checkbox
-      [matTooltip]="tooltip$ | async"
-      [matTooltipPosition]="tooltipPosition$ | async"
-      [checked]="cellValue$ | async"
-      [disabled]="disabled$ | async"
-      [indeterminate]="options.indeterminate ?? false"
-      [color]="options.theme ?? 'primary'"
+      [matTooltip]="tooltip()"
+      [matTooltipPosition]="tooltipPosition()"
+      [checked]="cellValue()"
+      [disabled]="disabled()"
+      [indeterminate]="renderOptions()?.indeterminate ?? false"
+      [color]="renderOptions()?.theme ?? 'primary'"
       (change)="onValueChange($event.checked)"
-      (click)="$event.stopImmediatePropagation()"
-    ></mat-checkbox>
-  }`,
+      (click)="$event.stopImmediatePropagation()" />
+  `,
 })
 export class CheckboxCellRendererComponent extends CellRendererAbstract<CheckboxCellRendererOptions> {
-  public readonly disabled$ = combineLatest([
-    this.columnConfig$,
-    this.data$,
-  ]).pipe(
-    map(
-      ([config, data]) => config?.cellEditorOptions?.disabled?.(data) ?? false,
-    ),
-  );
+  public readonly disabled = computed(() => this.columnConfig()?.cellEditorOptions?.disabled?.(this.data()) ?? false);
+
   public onValueChange(newValue: boolean): void {
-    this.columnConfig$
-      .pipe(take(1), withLatestFrom(this.data$))
-      .subscribe(([config, data]) => {
-        if (!this.handleValueChanged) {
-          throw Error(
-            `No handleValueChanged method provided for column ${config.key}`,
-          );
-        }
-        this.handleValueChanged?.(newValue, config, data);
-      });
+    const valueChanged = this.handleValueChanged();
+    if (!valueChanged) {
+      throw Error(`No handleValueChanged method provided for column ${this.columnConfig().key}`);
+    }
+    valueChanged(newValue, this.columnConfig(), this.data());
   }
 }
