@@ -1,35 +1,17 @@
-import { Observable, ReplaySubject } from 'rxjs';
 import { TableCell } from '../models/table-cell.model';
-import { Directive, Input } from '@angular/core';
-import { map, shareReplay } from 'rxjs/operators';
+import { computed, Directive, input } from '@angular/core';
 import { readPropValue } from '../../utils/utils';
 
 @Directive()
-export abstract class ColumnHeaderRendererAbstract<
-  ColumnHeaderRenderOptions = any,
-  T = any,
-> {
-  protected readonly _columnConfig$ = new ReplaySubject<TableCell<T>>();
-  public readonly columnConfig$: Observable<TableCell<T>> = this._columnConfig$
-    .asObservable()
-    .pipe(shareReplay({ bufferSize: 1, refCount: true }));
+export abstract class ColumnHeaderRendererAbstract<ColumnHeaderRenderOptions = any, T = any> {
+  public columnConfig = input.required<TableCell<T>>();
+  public disableSort = input<boolean>(false);
 
-  @Input({ required: true })
-  public set columnConfig(value: TableCell<T>) {
-    this._columnConfig$.next(value);
-  }
-
-  @Input()
-  public disableSort = false;
-
-  public readonly rendererOptions$: Observable<
-    ColumnHeaderRenderOptions | undefined
-  > = this.columnConfig$.pipe(
-    map((config) => config.headerRenderOptions),
-    shareReplay({ bufferSize: 1, refCount: true }),
+  public readonly renderOptions = computed<ColumnHeaderRenderOptions | undefined>(
+    () => this.columnConfig().headerRenderOptions
   );
 
-  public readonly columnLabel$ = this.columnConfig$.pipe(
-    map((c) => readPropValue<TableCell<T>>(c.label, c)),
+  public readonly columnLabel = computed(() =>
+    readPropValue<TableCell<T>>(this.columnConfig().label, this.columnConfig())
   );
 }
