@@ -1,9 +1,4 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  ViewEncapsulation,
-} from '@angular/core';
-import { AsyncPipe } from '@angular/common';
+import { ChangeDetectionStrategy, Component, ViewEncapsulation } from '@angular/core';
 import { ColumnHeaderRendererAbstract } from '../column-header-renderer-abstract.directive';
 import { ColumnHeaderSortingComponent } from '../column-header-sorting/column-header-sorting.component';
 import { ColumnHeaderWithIconRendererOptions } from './column-header-with-icon-renderer.options';
@@ -13,28 +8,20 @@ import { TranslateModule } from '@ngx-translate/core';
 @Component({
   selector: 'lab900-default-column-header-renderer',
   standalone: true,
-  imports: [
-    AsyncPipe,
-    ColumnHeaderSortingComponent,
-    MatIconModule,
-    TranslateModule,
-  ],
+  imports: [ColumnHeaderSortingComponent, MatIconModule, TranslateModule],
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None,
   template: `
-    @if (rendererOptions$ | async; as options) {
+    @if (renderOptions(); as options) {
       <div class="lab900-column-header lab900-column-header--with-icon">
-        @if (!options.svgIcon) {
+        @if (options.icon) {
           <mat-icon>{{ options.icon }}</mat-icon>
+        } @else if (options.svgIcon) {
+          <mat-icon [svgIcon]="options.icon" />
         }
-        @if (options.svgIcon) {
-          <mat-icon [svgIcon]="options.icon"></mat-icon>
-        }
-        {{ columnLabel$ | async | translate }}
+        {{ columnLabel() | translate }}
         @if (!disableSort) {
-          <lab900-column-header-sorting
-            [columnConfig]="columnConfig$ | async"
-          ></lab900-column-header-sorting>
+          <lab900-column-header-sorting [columnConfig]="columnConfig()" />
         }
       </div>
     }
