@@ -9,14 +9,12 @@ export abstract class SubscriptionBasedDirective implements OnDestroy {
     observable: Observable<T>,
     successCallback: (data: T) => void,
     errorCallback?: (error: any) => void,
-    completeCallback?: () => void,
+    completeCallback?: () => void
   ): void {
-    this.subscriptions.push(
-      observable.subscribe(successCallback, errorCallback, completeCallback),
-    );
+    this.subscriptions.push(observable.subscribe(successCallback, errorCallback, completeCallback));
   }
 
   public ngOnDestroy(): void {
-    this.subscriptions.forEach((sub) => sub.unsubscribe());
+    this.subscriptions.forEach(sub => sub.unsubscribe());
   }
 }
