@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
-import { Lab900ActionButtonComponent, Lab900ButtonComponent } from '@lab900/ui';
+import { Lab900ActionButtonComponent, Lab900ButtonComponent, ToggleActionButton } from '@lab900/ui';
+import { of } from 'rxjs';
 
 @Component({
   selector: 'lab900-button-example',
@@ -14,8 +15,7 @@ import { Lab900ActionButtonComponent, Lab900ButtonComponent } from '@lab900/ui';
       type="raised"
       label="hello world"
       suffixIcon="remove_red_eye"
-      throttleTimeInMs="1000"
-    />
+      throttleTimeInMs="1000" />
     <p>Raised button disabled</p>
     <lab900-button
       disabled="true"
@@ -23,16 +23,14 @@ import { Lab900ActionButtonComponent, Lab900ButtonComponent } from '@lab900/ui';
       color="primary"
       type="raised"
       label="hello world"
-      suffixIcon="remove_red_eye"
-    />
+      suffixIcon="remove_red_eye" />
     <p>Stroked button</p>
     <lab900-button
       (btnClick)="log('Stroked button')"
       color="primary"
       type="stroked"
       label="hello world"
-      prefixIcon="edit"
-    />
+      prefixIcon="edit" />
     <p>Stroked button disabled</p>
     <lab900-button
       (btnClick)="log('Stroked button disabled')"
@@ -40,30 +38,21 @@ import { Lab900ActionButtonComponent, Lab900ButtonComponent } from '@lab900/ui';
       color="primary"
       type="stroked"
       label="hello world"
-      prefixIcon="edit"
-    />
+      prefixIcon="edit" />
     <p>Raised button color= accent</p>
     <lab900-button color="accent" type="raised" label="hello world" />
     <p>Icon button</p>
     <lab900-button type="icon" label="delete" (btnClick)="log('Icon button')" />
     <p>Fab icon button</p>
-    <lab900-button
-      type="fab"
-      label="delete"
-      (btnClick)="log(' Fab Icon button')"
-    />
+    <lab900-button type="fab" label="delete" (btnClick)="log(' Fab Icon button')" />
     <p>Fab-mini icon button</p>
-    <lab900-button
-      type="mini-fab"
-      label="delete"
-      (btnClick)="log('Fab Mini Icon button')"
-    />
+    <lab900-button type="mini-fab" label="delete" (btnClick)="log('Fab Mini Icon button')" />
     <p>Toggle action button</p>
     <lab900-action-button [data]="[]" [action]="toggleActionButton" />
   `,
 })
 export class ButtonExampleComponent {
-  public toggleActionButton = {
+  public toggleActionButton: ToggleActionButton = {
     label: 'EntityDetailSectionToggleToggleButton',
     type: 'toggle',
     subActions: [
@@ -80,6 +69,13 @@ export class ButtonExampleComponent {
         action: () => {
           console.log('Button 2 is selected');
         },
+      },
+      {
+        label: 'Button 3',
+        action: () => {
+          console.log('Button 2 is selected');
+        },
+        hide: () => of(true),
       },
     ],
   };
