@@ -32,7 +32,7 @@ export const mergerSchemaExample: MergeConfig<MergerDataExample>[] = [
   {
     attribute: 'languages',
     label: 'language',
-    formatter: (data) => of(data.sort().join(', ')),
+    formatter: data => of(data.sort().join(', ')),
     nextLine: true,
     combine: true,
     active: true,
